CAUSEWAY-3815: remove behavior from enum  ActionResultResponseType
