+++ viewers/wicket/ui/src/main/java/org/apache/causeway/viewer/wicket/ui/components/widgets/actionlink/ActionLink.java
@@ -208,7 +208,7 @@ implements HasMetaModelContext, Menuable, HasManagedAction {
             var targetAdapter = actionModel.getParentObject();
 
             var bookmark = targetAdapter.refreshBookmark().orElseThrow();
-            getMetaModelContext().getTransactionService().flushTransaction();
+            getTransactionService().flushTransaction();
 
             // "redirect-after-post"
             RequestCycle.get().setResponsePage(EntityPage.class,
